+++ extensions/mongo/tests/test/MapperTestUtils.cpp
@@ -104,7 +104,7 @@ namespace catapult { namespace test {
 		EXPECT_EQ(block.Beneficiary, GetKeyValue(dbBlock, "beneficiary"));
 		EXPECT_EQ(block.FeeInterest, GetUint32(dbBlock, "feeInterest"));
 		EXPECT_EQ(block.FeeInterestDenominator, GetUint32(dbBlock, "feeInterestDenominator"));
-		EXPECT_EQ(block.Round, GetUint32(dbBlock, "round"));
+		EXPECT_EQ(block.Round, GetUint64(dbBlock, "round"));
 		EXPECT_EQ(block.CommitteePhaseTime, GetUint64(dbBlock, "committeePhaseTime"));
 		EXPECT_EQ(block.TransactionPayloadSize, GetUint32(dbBlock, "transactionPayloadSize"));
 	}
