+++ src/catapult/model/Block.h
@@ -71,7 +71,7 @@ namespace catapult { namespace model {
 		uint32_t FeeInterestDenominator;
 
 		/// Committee round (number of attempts to generate this block).
-		int16_t Round;
+		int64_t Round;
 
 		/// Time lapse in milliseconds of the committee phase of the round this block was generated in.
 		uint64_t CommitteePhaseTime;
